third_party/blink/web_tests/fast/table/large-col-span-crash.html

<!doctype html>
<style>
table {
  width: 100px;
  table-layout: fixed;
  height: 100px;
  background: lime;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<table data-expected-width=2004 data-expected-height=100>
  <col span="4294967295">
  <col>
</table>
<script>checkLayout("table");</script>
Pass if no crash, ensured by the checkLayout.
